Move simpleTrace definition back to Executive
